+++ examples/calendar.js
@@ -12,16 +12,7 @@ const formatter = new DateTimeFormat('yyyy-MM-dd HH:mm:ss');
 var Test = React.createClass({
   onChange(value) {
     console.log('DatePicker change: ' + (value && formatter.format(value)));
-  },
-
-  onCalendarSelect(value) {
-    console.log('calendar select: ' + (value && formatter.format(value)));
-    // controlled value
-    this.setState({
-      time: Date.now(),
-      open: this.state.showTime,
-      value: value
-    });
+    this.setState({value});
   },
 
   getInitialState() {
@@ -29,7 +20,7 @@ var Test = React.createClass({
     value.setTime(Date.now());
     return {
       open: false,
-      time: Date.now(),
+
       showTime: true,
       value: value
     };
@@ -44,10 +35,9 @@ var Test = React.createClass({
   render() {
     var state = this.state;
     var calendar = <Calendar locale={CalendarLocale}
-                             orient={['top', 'left']}
-                             showTime={this.state.showTime} onSelect={this.onCalendarSelect}
-                             onClear={this.onCalendarSelect.bind(this, null)} showClear={true}/>;
-    return <div className="form-group" style={{width: 400, margin: 20}} data-time={this.state.time}>
+                             showTime={this.state.showTime}
+                             showClear={true}/>;
+    return <div className="form-group" style={{width: 400, margin: 20}} >
       <div className="input-group">
         <span>
           <input type='checkbox' checked={this.state.showTime} onChange={this.onShowTimeChange}/>